Restructure project with nitra package
