+++ nitra/__init__.py
@@ -1,0 +1,1 @@
+"""Nitra terminal UI package."""
