--- panels/deployments_panel.py
@@ -1,5 +1,0 @@
-from textual.widgets import Static
-
-class DeploymentsPanel(Static):
-    def compose(self):
-        yield Static("Deployments Panel")
