--- panels/logs_panel.py
@@ -1,5 +1,0 @@
-from textual.widgets import Static
-
-class LogsPanel(Static):
-    def compose(self):
-        yield Static("Logs Panel")
