--- panels/portfolio_panel.py
@@ -1,5 +1,0 @@
-from textual.widgets import Static
-
-class PortfolioPanel(Static):
-    def compose(self):
-        yield Static("Portfolio Panel")
