--- panels/strategy_panel.py
@@ -1,5 +1,0 @@
-from textual.widgets import Static
-
-class StrategyPanel(Static):
-    def compose(self):
-        yield Static("Strategy Panel")
